Move errcheck package out of internal
